[Bug] adding tracks to lines doesnt work and fails (#28)
* Update wrong indices for LinienStreckenKonfigurationen

Modified unique constraints on LinienStreckenKonfigurationen to include VonBetriebspunktId and BisBetriebspunktId. This eliminates an error which was produced when a track is used mor than once, which is completely normal.

* Initialize `_betriebspunkte` on component load.

Added `OnInitializedAsync` to fetch and initialize `_betriebspunkte` when a `StreckenKonfiguration` is selected. This ensures the Betriebspunkte list is populated at the correct lifecycle stage.

* Refactor LoadFahrordnung component for consistency and functionality

Renamed private members with underscores for naming consistency. Added support for BetriebspunkteService and enhanced the initialization logic to load Betriebspunkte and transform line types. Improved select and validation logic for Linien and Strecken.

* Trim entries to match configuration bounds in FahrenRepository

Added logic to remove entries outside the defined 'VonBetriebspunktId' and 'BisBetriebspunktId' bounds in ListEntriesByStrecke results. This ensures that only the relevant entries for each configuration are included, maintaining data consistency.

* Fix method override placement in OnInitializedAsync

    Corrected the placement of the "override" modifier in the OnInitializedAsync method for better consistency and adherence to C# coding standards. This ensures proper method overriding and improved code readability.